Restructure files since `boat-swift5` wasn't recognized as a generator
